requirements.txt
async-timeout==4.0.2
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.0.4
colorlog==6.8.2
dataclasses==0.8
docopt==0.6.2
Flask==2.0.3
idna==3.7
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pipreqs==0.4.11
pycrypto==2.6.1
pyparsing==3.1.4
PyYAML==6.0.1
redis==4.3.6
requests==2.27.1
typing_extensions==4.1.1
urllib3==1.26.18
Werkzeug==2.0.3
yarg==0.1.9
zipp==3.6.0